Add .env file for image name
This will enable setting a .env file with IMAGE=username/pdi in it to
override the image name in both the spoon script and the makefile.
When building or running in the makefile it will use the custom image
name and when using the spoon script it will use the custom image name.
This way any forks from here out can set a .env file with their own repo
name instead of needing to hardcode their own username into the code.
It should also allow for CI scripts to build and publish containers a
bit more easily.

This patch does not change the default behavior of the makefile or spoon
script, however the makefile's default image name is pdi, and the spoon
scripts default image name is tgmti/pdi. This should maybe be
standardized to the same default value at some point.